+++ src/queue/processors/db/export-notes.ts
@@ -42,10 +42,9 @@ export async function exportNotes(job: Bull.Job, done: any): Promise<void> {
 	});
 
 	let exportedNotesCount = 0;
-	let ended = false;
 	let cursor: any = null;
 
-	while (!ended) {
+	while (true) {
 		const notes = await Note.find({
 			userId: user._id,
 			...(cursor ? { _id: { $gt: cursor } } : {})
@@ -57,7 +56,6 @@ export async function exportNotes(job: Bull.Job, done: any): Promise<void> {
 		});
 
 		if (notes.length === 0) {
-			ended = true;
 			job.progress(100);
 			break;
 		}
@@ -101,7 +99,7 @@ export async function exportNotes(job: Bull.Job, done: any): Promise<void> {
 	logger.succ(`Exported to: ${path}`);
 
 	const fileName = 'notes-' + dateFormat(new Date(), 'yyyy-mm-dd-HH-MM-ss') + '.json';
-	const driveFile = await addFile(user, path, fileName);
+	const driveFile = await addFile(user, path, fileName, null, null, true);
 
 	logger.succ(`Exported to: ${driveFile._id}`);
 	cleanup();