+++ src/queue/processors/inbox.ts
@@ -11,6 +11,8 @@ import Logger from '../../services/logger';
 import { registerOrFetchInstanceDoc } from '../../services/register-or-fetch-instance-doc';
 import Instance from '../../models/instance';
 import instanceChart from '../../services/chart/instance';
+import { validActor } from '../../remote/activitypub/type';
+import { toDbHost } from '../../misc/convert-host';
 
 const logger = new Logger('inbox');
 
@@ -46,7 +48,7 @@ export default async (job: Bull.Job): Promise<void> => {
 
 		// ブロックしてたら中断
 		// TODO: いちいちデータベースにアクセスするのはコスト高そうなのでどっかにキャッシュしておく
-		const instance = await Instance.findOne({ host: host.toLowerCase() });
+		const instance = await Instance.findOne({ host: toDbHost(host) });
 		if (instance && instance.isBlocked) {
 			logger.info(`Blocked request: ${host}`);
 			return;
@@ -65,7 +67,7 @@ export default async (job: Bull.Job): Promise<void> => {
 
 		// ブロックしてたら中断
 		// TODO: いちいちデータベースにアクセスするのはコスト高そうなのでどっかにキャッシュしておく
-		const instance = await Instance.findOne({ host: host.toLowerCase() });
+		const instance = await Instance.findOne({ host: toDbHost(host) });
 		if (instance && instance.isBlocked) {
 			logger.warn(`Blocked request: ${host}`);
 			return;
@@ -79,7 +81,7 @@ export default async (job: Bull.Job): Promise<void> => {
 
 	// Update Person activityの場合は、ここで署名検証/更新処理まで実施して終了
 	if (activity.type === 'Update') {
-		if (activity.object && activity.object.type === 'Person') {
+		if (activity.object && validActor.includes(activity.object.type)) {
 			if (user == null) {
 				logger.warn('Update activity received, but user not registed.');
 			} else if (!httpSignature.verifySignature(signature, user.publicKey.publicKeyPem)) {