+++ src/server/api/endpoints/ap/show.ts
@@ -10,6 +10,7 @@ import Resolver from '../../../../remote/activitypub/resolver';
 import { ApiError } from '../../error';
 import Instance from '../../../../models/instance';
 import { extractDbHost } from '../../../../misc/convert-host';
+import { validActor } from '../../../../remote/activitypub/type';
 
 export const meta = {
 	tags: ['federation'],
@@ -85,6 +86,17 @@ async function fetchAny(uri: string) {
 	// /@user のような正規id以外で取得できるURIが指定されていた場合、ここで初めて正規URIが確定する
 	// これはDBに存在する可能性があるため再度DB検索
 	if (uri !== object.id) {
+		if (object.id.startsWith(config.url + '/')) {
+			const id = new mongo.ObjectID(object.id.split('/').pop());
+			const [user, note] = await Promise.all([
+				User.findOne({ _id: id }),
+				Note.findOne({ _id: id })
+			]);
+
+			const packed = await mergePack(user, note);
+			if (packed !== null) return packed;
+		}
+
 		const [user, note] = await Promise.all([
 			User.findOne({ uri: object.id }),
 			Note.findOne({ uri: object.id })
@@ -95,7 +107,7 @@ async function fetchAny(uri: string) {
 	}
 
 	// それでもみつからなければ新規であるため登録
-	if (object.type === 'Person') {
+	if (validActor.includes(object.type)) {
 		const user = await createPerson(object.id);
 		return {
 			type: 'User',
@@ -104,7 +116,7 @@ async function fetchAny(uri: string) {
 	}
 
 	if (['Note', 'Question', 'Article'].includes(object.type)) {
-		const note = await createNote(object.id);
+		const note = await createNote(object.id, null, true);
 		return {
 			type: 'Note',
 			object: await packNote(note, null, { detail: true })